Trait xor_fee::pallet::Config

pub trait Config: Config + Config + Config {

    type PermittedSetPeriod: EnsureOrigin<Self::RuntimeOrigin>;
    type DynamicMultiplier: CalculateMultiplier<AssetIdOf<Self>, DispatchError>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type XorCurrency: Currency<Self::AccountId> + Send + Sync;
    type XorId: Get<AssetIdOf<Self>>;
    type ValId: Get<AssetIdOf<Self>>;
    type KusdId: Get<AssetIdOf<Self>>;
    type TbcdId: Get<AssetIdOf<Self>>;
    type FeeReferrerWeight: Get<u32>;
    type FeeXorBurnedWeight: Get<u32>;
    type FeeValBurnedWeight: Get<u32>;
    type FeeKusdBurnedWeight: Get<u32>;
    type RemintTbcdBuyBackPercent: Get<Percent>;
    type RemintKusdBuyBackPercent: Get<Percent>;
    type DEXIdValue: Get<Self::DEXId>;
    type LiquidityProxy: LiquidityProxyTrait<Self::DEXId, Self::AccountId, AssetIdOf<Self>>;
    type OnValBurned: OnValBurned;
    type CustomFees: ApplyCustomFees<<Self as Config>::RuntimeCall, Self::AccountId>;
    type GetTechnicalAccountId: Get<Self::AccountId>;
    type FullIdentification;
    type ReferrerAccountProvider: ReferrerAccountProvider<Self::AccountId>;
    type BuyBackHandler: BuyBackHandler<Self::AccountId, AssetIdOf<Self>>;
    type WeightInfo: WeightInfo;
    type WithdrawFee: WithdrawFee<Self>;
    type MaxWhiteListTokens: Get<u32>;
    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + GetDispatchInfo + From<Call<Self>>;
    type PoolXyk: LiquiditySource<Self::DEXId, Self::AccountId, AssetIdOf<Self>, Balance, DispatchError>;
    type WhiteListOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type PriceTools: PriceToolsProvider<AssetIdOf<Self>>;
    type MinimalFeeInAsset: Get<Balance>;
    type Randomness: Randomness<Self::Hash, Self::BlockNumber>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type PermittedSetPeriod: EnsureOrigin<Self::RuntimeOrigin>

type DynamicMultiplier: CalculateMultiplier<AssetIdOf<Self>, DispatchError>

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type XorCurrency: Currency<Self::AccountId> + Send + Sync

XOR - The native currency of this blockchain.

type XorId: Get<AssetIdOf<Self>>

type ValId: Get<AssetIdOf<Self>>

type KusdId: Get<AssetIdOf<Self>>

type TbcdId: Get<AssetIdOf<Self>>

type FeeReferrerWeight: Get<u32>

type FeeXorBurnedWeight: Get<u32>

type FeeValBurnedWeight: Get<u32>

type FeeKusdBurnedWeight: Get<u32>

type RemintTbcdBuyBackPercent: Get<Percent>

type RemintKusdBuyBackPercent: Get<Percent>

type DEXIdValue: Get<Self::DEXId>

type LiquidityProxy: LiquidityProxyTrait<Self::DEXId, Self::AccountId, AssetIdOf<Self>>

type OnValBurned: OnValBurned

type CustomFees: ApplyCustomFees<<Self as Config>::RuntimeCall, Self::AccountId>

type GetTechnicalAccountId: Get<Self::AccountId>

type FullIdentification

type ReferrerAccountProvider: ReferrerAccountProvider<Self::AccountId>

type BuyBackHandler: BuyBackHandler<Self::AccountId, AssetIdOf<Self>>

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type WithdrawFee: WithdrawFee<Self>

type MaxWhiteListTokens: Get<u32>

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + GetDispatchInfo + From<Call<Self>>

type PoolXyk: LiquiditySource<Self::DEXId, Self::AccountId, AssetIdOf<Self>, Balance, DispatchError>

type WhiteListOrigin: EnsureOrigin<Self::RuntimeOrigin>

type PriceTools: PriceToolsProvider<AssetIdOf<Self>>

type MinimalFeeInAsset: Get<Balance>

Main goal of the constant is to prevent zero fees

type Randomness: Randomness<Self::Hash, Self::BlockNumber>